move status condition definitions to correct mod definition
